minor optimizations

+++ src/main/java/com/springboot/tinyurlspringboot/controller/ShortenerController.java
@@ -26,33 +26,40 @@ public class ShortenerController {
     public ResponseEntity<String> shortener(@RequestBody Map<String,String> map) {
         String email = map.get("email");
         String original = map.get("original");
-        original = URLValidator.normalize(original);
         String shortURL = "";
         String urlName=map.get("urlName");
-        User user=userRepository.findByEmail(email);
-        Shortener existing = shortenerRepository.findByOriginalAndUser(original,user);
-        if (user==null){
-            return ResponseEntity.status(404).body("User not found");
-        }
         if (!URLValidator.isValidURL(original)){
             return ResponseEntity.badRequest().body("Invalid URL");
         }
-        if (existing==null) {
-            Shortener newShortener = shortenerRepository.save(new Shortener(original,shortURL,urlName,user));
-            long id = newShortener.getId();
-            shortURL = ShortenerService.getShortURL(id+1000000);
-            newShortener.setShort(shortURL);
-            shortenerRepository.save(newShortener);
-            return ResponseEntity.ok(shortURL);
+        User user=userRepository.findByEmail(email);
+        if (user==null){
+            return ResponseEntity.status(404).body("User not found");
         }
-        return ResponseEntity.ok(existing.getShortUrl());
+        original = URLValidator.normalize(original);
+        Shortener existing = shortenerRepository.findByOriginalAndUser(original,user);
+        if (existing != null) {
+            return ResponseEntity.ok(existing.getShortUrl());
+        }
+
+        Long nextId = shortenerRepository.getNextSeriesId();
+        String generatedShortUrl = ShortenerService.getShortURL(nextId);
+        Shortener newShortener = new Shortener(original, generatedShortUrl, urlName, user);
+        newShortener.setId(nextId);
+        shortenerRepository.save(newShortener);
+
+        return ResponseEntity.ok(generatedShortUrl);
+
     }
+
     @PostMapping("/delete")
     public void delete(@RequestBody Map<String,String> map) {
-        String email=map.get("email");
-        String shortURL=map.get("shortURL");
-        User user=userRepository.findByEmail(email);
-        Shortener shortener=shortenerRepository.findByShortUrlAndUser(shortURL,user);
-        shortenerRepository.delete(shortener);
-    }
-}
+        String email = map.get("email");
+        String shortURL = map.get("shortURL");
+        User user = userRepository.findByEmail(email);
+        if (user != null) {
+            Shortener shortener = shortenerRepository.findByShortUrlAndUser(shortURL, user);
+            if (shortener != null) {
+                shortenerRepository.delete(shortener);
+            }
+        }
+    }}
